+++ src/main/java/duke/command/Command.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.common.DukeException;
+import duke.storage.Storage;
+import duke.task.TaskList;
+import duke.ui.Ui;
+
 public abstract class Command {
     public abstract void execute(Ui ui, Storage storage, TaskList taskList) throws DukeException;
