+++ src/main/java/duke/common/DukeException.java
@@ -1,0 +1,7 @@
+package duke.common;
+
+public class DukeException extends Exception {
+    public DukeException(String message) {
+        super(message);
+    }
+}
