+++ src/main/java/duke/task/Deadline.java
@@ -1,6 +1,9 @@
+package duke.task;
+
+import duke.common.DukeException;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
-import java.util.Arrays;
 
 public class Deadline extends Task {
     private LocalDate localDate;
