+++ src/main/java/duke/task/ToDo.java
@@ -1,3 +1,7 @@
+package duke.task;
+
+import duke.common.DukeException;
+
 public class ToDo extends Task {
     public ToDo(String description, boolean isDone) throws DukeException {
         super("todo", description, isDone);
