+++ text-ui-test/runtest.bat
@@ -10,7 +10,15 @@ REM delete save file from previous run
 if exist data\tasks.txt del data\tasks.txt
 
 REM compile the code into the bin folder
-javac  -cp ..\src\main\java -Xlint:none -d ..\bin ..\src\main\java\*.java
+javac  -cp ..\src\main\java\duke -Xlint:none -d ..\bin ^
+..\src\main\java\duke\*.java ^
+..\src\main\java\duke\command\*.java ^
+..\src\main\java\duke\common\*.java ^
+..\src\main\java\duke\parser\*.java ^
+..\src\main\java\duke\storage\*.java ^
+..\src\main\java\duke\task\*.java ^
+..\src\main\java\duke\ui\*.java
+
 IF ERRORLEVEL 1 (
     echo ********** BUILD FAILURE **********
     exit /b 1
@@ -18,7 +26,7 @@ IF ERRORLEVEL 1 (
 REM no error here, errorlevel == 0
 
 REM run the program, feed commands from input.txt file and redirect the output to the ACTUAL.TXT
-java -classpath ..\bin Duke < input.txt > ACTUAL.TXT
+java -classpath ..\bin duke.Duke < input.txt > ACTUAL.TXT
 
 REM compare the output to the expected output
 FC ACTUAL.TXT EXPECTED.TXT
